navbar-ui (#1)
In this pull request, I organized the basic structure of my navbar and set up an initial draft related to the styles, which will be reviewed when I apply the color palette I'll be using. I also created a new button as a separate component, which is imported to enhance code cleanliness. Similarly, I use icons that allow the user to navigate to my professional social networks.